chore: *.js -> *.mjs
